[HW01] Replaced regexp by trim
